More Cursed Parsing
Further improved parsing and did a lot more with the ast generation (and still a lot more to do). Also outlined the interpretor/vm in another file so hopefully some work can begin on that (some ideas for future resolutions for awaits and next references and negative lifetimes are in the works).